backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
cloudinary>=1.36
pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.0
psycopg2-binary>=2.9
